+++ src/app/login/login.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {AuthService} from "../auth.service";
+import {AuthService} from "../../utils/auth.service";
 import {FormControl, FormGroup} from "@angular/forms";
 import {Router} from "@angular/router";
