+++ src/app/pollution/pollution.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import {CommuneService} from "../../utils/commune.service";
+import {AuthService} from "../../utils/auth.service";
+import {Observable} from "rxjs";
 
 @Component({
   selector: 'app-pollution',
@@ -6,10 +9,14 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./pollution.component.scss']
 })
 export class PollutionComponent implements OnInit {
+  auth_token = '';
+  communesArr: any = [];
 
-  constructor() { }
+  constructor(private authS: AuthService, private communeS: CommuneService) { }
 
   ngOnInit(): void {
   }
-
 }
+
+
+
