+++ src/app/register/register.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {FormControl, FormGroup} from "@angular/forms";
-import {AuthService} from "../auth.service";
+import {AuthService} from "../../utils/auth.service";
 import {Router} from "@angular/router";
 
 @Component({
